Don't download or use the enhanced client when uninstalling everything
The new uninstall functionality will use the Helm chart to uninstall
the traffic-manager and the `telepresence-agents` config map to
uninstall traffic-agents. It never talks to the traffic-manager and
consequently, have no use of being logged in.

This commit ensures that the user is logged out before an `uninstall
--everything` starts, so that the command doesn't make an attempt to
download or upgrade the enhanced free client.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>